feat(desktop): add workspace tabs and AWS auth
