+++ desktop/frontend/src/style.css
@@ -11,3 +11,29 @@ body {
   background-color: #0f172a; /* slate-900 */
   color: #f1f5f9; /* slate-100 */
 }
+
+.workspaces-container {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+}
+
+.tabs {
+  display: flex;
+  background-color: #1e293b;
+}
+
+.tab {
+  padding: 0.5rem 1rem;
+  cursor: pointer;
+  border-bottom: 2px solid transparent;
+}
+
+.tab.active {
+  border-bottom-color: #38bdf8;
+}
+
+.tab.add {
+  padding: 0.5rem;
+}
+
